modularize sql generator
