+++ generators/sql/generator.mts
@@ -1,0 +1,68 @@
+/**
+ * Copyright 2013-2023 the original author or authors from the JHipster project.
+ *
+ * This file is part of the JHipster project, see https://www.jhipster.tech/
+ * for more information.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+import _ from 'lodash';
+
+import BaseApplicationGenerator from '../base-application/index.mjs';
+import { GENERATOR_SQL, GENERATOR_BOOTSTRAP_APPLICATION, GENERATOR_LIQUIBASE } from '../generator-list.mjs';
+import writeTask from './files.mjs';
+import cleanupTask from './cleanup.mjs';
+import writeEntitiesTask, { cleanupEntitiesTask } from './entity-files.mjs';
+
+export default class SqlGenerator extends BaseApplicationGenerator {
+  async beforeQueue() {
+    await this.dependsOnJHipster(GENERATOR_BOOTSTRAP_APPLICATION);
+    if (!this.fromBlueprint) {
+      await this.composeWithBlueprints(GENERATOR_SQL);
+    }
+  }
+
+  get composing() {
+    return this.asComposingTaskGroup({
+      async composing() {
+        await this.composeWithJHipster(GENERATOR_LIQUIBASE);
+      },
+    });
+  }
+
+  get [BaseApplicationGenerator.COMPOSING]() {
+    return this.asComposingTaskGroup(this.delegateTasksToBlueprint(() => this.composing));
+  }
+
+  get writing() {
+    return this.asWritingTaskGroup({
+      cleanupTask,
+      writeTask,
+    });
+  }
+
+  get [BaseApplicationGenerator.WRITING]() {
+    return this.delegateTasksToBlueprint(() => this.writing);
+  }
+
+  get writingEntities() {
+    return {
+      cleanupEntitiesTask,
+      writeEntitiesTask,
+    };
+  }
+
+  get [BaseApplicationGenerator.WRITING_ENTITIES]() {
+    return this.delegateTasksToBlueprint(() => this.writingEntities);
+  }
+}
