+++ generators/sql/generator.spec.mts
@@ -1,31 +1,34 @@
 import { jestExpect as expect } from 'mocha-expect-snapshot';
-import { dirname, join } from 'path';
+import lodash from 'lodash';
+import { basename, dirname, join } from 'path';
 import { fileURLToPath } from 'url';
 
-import {
-  buildServerMatrix,
-  extendMatrix,
-  extendFilteredMatrix,
-  entitiesServerSamples as entities,
-  buildSamplesFromMatrix,
-} from '../../test/support/index.mjs';
+import { buildServerMatrix, extendMatrix, extendFilteredMatrix, buildSamplesFromMatrix } from '../../test/support/index.mjs';
+import { testBlueprintSupport } from '../../test/support/tests.mjs';
+import Generator from '../server/index.mjs';
 import { defaultHelpers as helpers } from '../../test/support/helpers.mjs';
 
 import { databaseTypes, cacheTypes } from '../../jdl/jhipster/index.mjs';
+import { mockedGenerators, shouldComposeWithKafka, shouldComposeWithLiquibase } from '../server/__test-support/index.mjs';
+import { GENERATOR_SERVER } from '../generator-list.mjs';
+
+const { snakeCase } = lodash;
 
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = dirname(__filename);
 
+const generator = basename(__dirname);
 const generatorFile = join(__dirname, 'index.mjs');
 
 const { SQL: databaseType, H2_DISK, H2_MEMORY, POSTGRESQL, MARIADB, MYSQL, MSSQL, ORACLE } = databaseTypes;
 const commonConfig = { databaseType, baseName: 'jhipster', nativeLanguage: 'en', languages: ['fr', 'en'] };
 const { NO: NO_CACHE_PROVIDER, EHCACHE, CAFFEINE, HAZELCAST, INFINISPAN, MEMCACHED, REDIS } = cacheTypes;
 
-let sqlSamples = buildServerMatrix();
+let sqlSamples = buildServerMatrix({
+  prodDatabaseType: [POSTGRESQL, MARIADB, MYSQL, MSSQL, ORACLE],
+});
 
 sqlSamples = extendMatrix(sqlSamples, {
-  prodDatabaseType: [POSTGRESQL, MARIADB, MYSQL, MSSQL, ORACLE],
   enableHibernateCache: [false, true],
 });
 
@@ -54,25 +57,29 @@ sqlSamples = extendFilteredMatrix(sqlSamples, ({ reactive }) => !reactive, {
 });
 
 const testSamples = buildSamplesFromMatrix(sqlSamples, { commonConfig });
-const skipPriorities = ['writing', 'postWriting'];
 
-describe(`generator - ${databaseType} - entities`, () => {
+describe(`generator - ${databaseType}`, () => {
+  it('generator-list constant matches folder name', async () => {
+    await expect((await import('../generator-list.mjs'))[`GENERATOR_${snakeCase(generator).toUpperCase()}`]).toBe(generator);
+  });
+  it('should support features parameter', () => {
+    const instance = new Generator([], { help: true, env: { cwd: 'foo', sharedOptions: { sharedData: {} } } }, { unique: 'bar' });
+    expect(instance.features.unique).toBe('bar');
+  });
+  describe('blueprint support', () => testBlueprintSupport(generator));
+
   it('samples matrix should match snapshot', () => {
     expect(testSamples).toMatchSnapshot();
   });
 
   Object.entries(testSamples).forEach(([name, sampleConfig]) => {
-    const { enableTranslation } = sampleConfig;
+    const { authenticationType, enableTranslation } = sampleConfig;
 
     describe(name, () => {
       let runResult;
 
       before(async () => {
-        runResult = await helpers
-          .run(generatorFile)
-          .withJHipsterConfig(sampleConfig, entities)
-          .withOptions({ skipPriorities })
-          .withMockedGenerators(['jhipster:languages', 'jhipster:common', 'jhipster:liquibase']);
+        runResult = await helpers.runJHipster(GENERATOR_SERVER).withJHipsterConfig(sampleConfig).withMockedGenerators(mockedGenerators);
       });
 
       after(() => runResult.cleanup());
@@ -89,6 +96,14 @@ describe(`generator - ${databaseType} - entities`, () => {
       it('should match generated files snapshot', () => {
         expect(runResult.getStateSnapshot()).toMatchSnapshot();
       });
+      it('contains correct authenticationType', () => {
+        runResult.assertFileContent('.yo-rc.json', new RegExp(`"authenticationType": "${authenticationType}"`));
+      });
+      it('contains correct databaseType', () => {
+        runResult.assertFileContent('.yo-rc.json', new RegExp(`"databaseType": "${databaseType}"`));
+      });
+      shouldComposeWithKafka(sampleConfig, () => runResult);
+      shouldComposeWithLiquibase(sampleConfig, () => runResult);
     });
   });
 });
