+++ generators/sql/templates/sql/common/src/main/java/package/config/DatabaseConfiguration.java.ejs
