+++ generators/sql/templates/sql/common/src/main/resources/.h2.server.properties.ejs
