+++ generators/sql/templates/sql/common/src/test/java/package/config/EmbeddedSQL.java.ejs
