+++ generators/sql/templates/sql/common/src/test/java/package/config/MariadbTestContainer.java.ejs
