+++ generators/sql/templates/sql/common/src/test/java/package/config/MysqlTestContainer.java.ejs
