+++ generators/sql/templates/sql/common/src/test/java/package/config/SqlTestContainer.java.ejs
