+++ generators/sql/templates/sql/common/src/test/java/package/config/timezone/HibernateTimeZoneIT.java.ejs
