+++ generators/sql/templates/sql/common/src/test/resources/config/application-testdev.yml.ejs
