+++ generators/sql/templates/sql/common/src/test/resources/config/application-testprod.yml.ejs
