+++ generators/sql/templates/sql/common/src/test/resources/testcontainers/mariadb/my.cnf
