+++ generators/sql/templates/sql/reactive/src/main/java/package/repository/EntityManager.java.ejs
