+++ generators/sql/templates/sql/reactive/src/main/java/package/repository/UserSqlHelper.java.ejs
