+++ generators/sql/templates/sql/reactive/src/main/java/package/repository/rowmapper/UserRowMapper.java.ejs
