+++ generators/sql/templates/src/main/java/package/domain/_PersistClass_.java.jhi.hibernate_cache.ejs
