+++ generators/sql/templates/src/main/java/package/domain/_PersistClass_.java.jhi.jakarta_lifecycle_events.ejs
