+++ generators/sql/templates/src/main/java/package/domain/_PersistClass_.java.jhi.jakarta_persistence.ejs
