+++ generators/sql/templates/src/main/java/package/domain/_PersistClass_.java.jhi.spring_data_reactive.ejs
