+++ generators/sql/templates/src/main/java/package/domain/_PersistClass_Callback.java.ejs
