+++ generators/sql/templates/src/main/java/package/repository/_EntityClass_RepositoryWithBagRelationshipsImpl.java.ejs
